test(db): move browse + author_photos_data + palette tests
