+++ db/src/palette.rs
@@ -333,3 +333,1009 @@ pub async fn search_palette(
         duration_ms,
     })
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::sync::test_helpers::indexed;
+    use crate::books::list_books;
+    use crate::covers::test_helpers::CoversTempDir;
+    use crate::metadata_overrides::upsert_metadata_overrides;
+    use crate::pool::init_db;
+    use crate::sync::replace_books;
+    use omnibus_shared::{Contributor, MetadataOverrides};
+    use sqlx::SqlitePool;
+
+    // ── search_palette ──────────────────────────────────────────────
+    #[tokio::test]
+    async fn palette_books_match_title() {
+        let _covers = CoversTempDir::new("palette_books");
+        let pool = init_db("sqlite::memory:").await.unwrap();
+        replace_books(
+            &pool,
+            "/lib",
+            vec![
+                indexed(
+                    "a.epub",
+                    Some("Dracula"),
+                    &["Bram Stoker"],
+                    &["Horror"],
+                    None,
+                    None,
+                ),
+                indexed(
+                    "b.epub",
+                    Some("Frankenstein"),
+                    &["Mary Shelley"],
+                    &["Horror"],
+                    None,
+                    None,
+                ),
+            ],
+        )
+        .await
+        .unwrap();
+
+        let results = search_palette(&pool, "/lib", "dracula").await.unwrap();
+        assert_eq!(results.books.len(), 1);
+        assert_eq!(results.books[0].title, "Dracula");
+        assert_eq!(results.books[0].author_display, "Bram Stoker");
+        assert!(results.books[0].formats.contains(&"EPUB".to_string()));
+    }
+    #[tokio::test]
+    async fn palette_authors_match_substring() {
+        let _covers = CoversTempDir::new("palette_authors");
+        let pool = init_db("sqlite::memory:").await.unwrap();
+        replace_books(
+            &pool,
+            "/lib",
+            vec![indexed(
+                "a.epub",
+                Some("Babel"),
+                &["R. F. Kuang"],
+                &[],
+                None,
+                None,
+            )],
+        )
+        .await
+        .unwrap();
+
+        let results = search_palette(&pool, "/lib", "kuang").await.unwrap();
+        assert!(!results.authors.is_empty(), "should match author substring");
+        assert_eq!(results.authors[0].name, "R. F. Kuang");
+        assert_eq!(results.authors[0].book_count, 1);
+    }
+    #[tokio::test]
+    async fn palette_series_match() {
+        let _covers = CoversTempDir::new("palette_series");
+        let pool = init_db("sqlite::memory:").await.unwrap();
+        replace_books(
+            &pool,
+            "/lib",
+            vec![
+                indexed(
+                    "a.epub",
+                    Some("Book One"),
+                    &["Author"],
+                    &[],
+                    Some(("Poppy War", "1")),
+                    None,
+                ),
+                indexed(
+                    "b.epub",
+                    Some("Book Two"),
+                    &["Author"],
+                    &[],
+                    Some(("Poppy War", "2")),
+                    None,
+                ),
+            ],
+        )
+        .await
+        .unwrap();
+
+        let results = search_palette(&pool, "/lib", "poppy").await.unwrap();
+        assert!(!results.series.is_empty(), "should match series substring");
+        assert_eq!(results.series[0].name, "Poppy War");
+        assert_eq!(results.series[0].book_count, 2);
+    }
+    #[tokio::test]
+    async fn palette_tags_match() {
+        let _covers = CoversTempDir::new("palette_tags");
+        let pool = init_db("sqlite::memory:").await.unwrap();
+        replace_books(
+            &pool,
+            "/lib",
+            vec![indexed(
+                "a.epub",
+                Some("A"),
+                &["Author"],
+                &["Dark academia"],
+                None,
+                None,
+            )],
+        )
+        .await
+        .unwrap();
+
+        let results = search_palette(&pool, "/lib", "academia").await.unwrap();
+        assert!(!results.tags.is_empty(), "should match tag substring");
+        assert_eq!(results.tags[0].name, "Dark academia");
+        assert_eq!(results.tags[0].book_count, 1);
+    }
+    /// Bug #1 (display side): the palette must show the overridden title,
+    /// not the canonical scanned `b.title`, so what the user clicks matches
+    /// what they searched for.
+    #[tokio::test]
+    async fn palette_book_hit_uses_overridden_title() {
+        let _covers = CoversTempDir::new("palette_override_title");
+        let pool = init_db("sqlite::memory:").await.unwrap();
+        let user_id = crate::auth::create_user(&pool, "admin", "securepassword1")
+            .await
+            .unwrap()
+            .id;
+
+        replace_books(
+            &pool,
+            "/lib",
+            vec![indexed(
+                "p.epub",
+                Some("Scanned Title"),
+                &["Author"],
+                &[],
+                None,
+                None,
+            )],
+        )
+        .await
+        .unwrap();
+        let uuid = list_books(&pool, "/lib").await.unwrap()[0]
+            .unique_identifier
+            .clone()
+            .unwrap();
+        upsert_metadata_overrides(
+            &pool,
+            &uuid,
+            &MetadataOverrides {
+                title: Some("Edited Title".into()),
+                ..Default::default()
+            },
+            false,
+            user_id,
+        )
+        .await
+        .unwrap();
+
+        let palette = search_palette(&pool, "/lib", "Edited").await.unwrap();
+        assert_eq!(palette.books.len(), 1);
+        assert_eq!(palette.books[0].title, "Edited Title");
+    }
+    /// Bug #1 (display side): overriding the creators list rebuilds the
+    /// comma-joined `author_display` so the palette subtitle matches the
+    /// detail page.
+    #[tokio::test]
+    async fn palette_book_hit_uses_overridden_author_display() {
+        let _covers = CoversTempDir::new("palette_override_authors");
+        let pool = init_db("sqlite::memory:").await.unwrap();
+        let user_id = crate::auth::create_user(&pool, "admin", "securepassword1")
+            .await
+            .unwrap()
+            .id;
+
+        replace_books(
+            &pool,
+            "/lib",
+            vec![indexed(
+                "p.epub",
+                Some("Searchable"),
+                &["Original Author"],
+                &[],
+                None,
+                None,
+            )],
+        )
+        .await
+        .unwrap();
+        let uuid = list_books(&pool, "/lib").await.unwrap()[0]
+            .unique_identifier
+            .clone()
+            .unwrap();
+        upsert_metadata_overrides(
+            &pool,
+            &uuid,
+            &MetadataOverrides {
+                creators: Some(vec![
+                    Contributor {
+                        name: "First Override".into(),
+                        ..Default::default()
+                    },
+                    Contributor {
+                        name: "Second Override".into(),
+                        ..Default::default()
+                    },
+                ]),
+                ..Default::default()
+            },
+            false,
+            user_id,
+        )
+        .await
+        .unwrap();
+
+        let palette = search_palette(&pool, "/lib", "Searchable").await.unwrap();
+        assert_eq!(palette.books.len(), 1);
+        assert_eq!(
+            palette.books[0].author_display,
+            "First Override, Second Override"
+        );
+    }
+    /// Palette book hits should surface a user-uploaded cover even when the
+    /// scanned book had no cover. Mirrors `apply_overrides` so the palette
+    /// row doesn't go cover-less for an override-only cover.
+    #[tokio::test]
+    async fn palette_book_hit_uses_overridden_cover() {
+        let _covers = CoversTempDir::new("palette_override_cover");
+        let pool = init_db("sqlite::memory:").await.unwrap();
+        let user_id = crate::auth::create_user(&pool, "admin", "securepassword1")
+            .await
+            .unwrap()
+            .id;
+
+        // Indexed book with no scanned cover.
+        replace_books(
+            &pool,
+            "/lib",
+            vec![indexed(
+                "p.epub",
+                Some("Coverless Searchable"),
+                &["Author"],
+                &[],
+                None,
+                None,
+            )],
+        )
+        .await
+        .unwrap();
+        let book = list_books(&pool, "/lib").await.unwrap().remove(0);
+        let uuid = book.unique_identifier.clone().unwrap();
+
+        // Set has_cover_override = true with no text edits.
+        upsert_metadata_overrides(&pool, &uuid, &MetadataOverrides::default(), true, user_id)
+            .await
+            .unwrap();
+
+        let palette = search_palette(&pool, "/lib", "Coverless").await.unwrap();
+        assert_eq!(palette.books.len(), 1);
+        assert_eq!(
+            palette.books[0].cover_url,
+            Some(format!("/api/covers/{uuid}"))
+        );
+    }
+    // #128: lock the wiring between the palette and `build_fts_match`'s
+    // facet prefixes. A regression in the facet parser could otherwise
+    // silently break palette tag:/author:/series: queries without any
+    // palette test failing.
+    #[tokio::test]
+    async fn palette_book_matches_tag_facet() {
+        let _covers = CoversTempDir::new("palette_tag_facet");
+        let pool = init_db("sqlite::memory:").await.unwrap();
+        replace_books(
+            &pool,
+            "/lib",
+            vec![
+                indexed(
+                    "a.epub",
+                    Some("Dracula"),
+                    &["Bram Stoker"],
+                    &["vampires"],
+                    None,
+                    None,
+                ),
+                indexed(
+                    "b.epub",
+                    Some("Frankenstein"),
+                    &["Mary Shelley"],
+                    &["monsters"],
+                    None,
+                    None,
+                ),
+            ],
+        )
+        .await
+        .unwrap();
+
+        let results = search_palette(&pool, "/lib", "tag:vampires").await.unwrap();
+        let titles: Vec<&str> = results.books.iter().map(|b| b.title.as_str()).collect();
+        assert!(
+            titles.contains(&"Dracula"),
+            "tag:vampires should match Dracula, got {titles:?}"
+        );
+        assert!(
+            !titles.contains(&"Frankenstein"),
+            "tag:vampires should not match Frankenstein"
+        );
+    }
+    #[tokio::test]
+    async fn palette_book_matches_author_facet() {
+        let _covers = CoversTempDir::new("palette_author_facet");
+        let pool = init_db("sqlite::memory:").await.unwrap();
+        replace_books(
+            &pool,
+            "/lib",
+            vec![
+                indexed(
+                    "a.epub",
+                    Some("Dracula"),
+                    &["Bram Stoker"],
+                    &["horror"],
+                    None,
+                    None,
+                ),
+                indexed(
+                    "b.epub",
+                    Some("Frankenstein"),
+                    &["Mary Shelley"],
+                    &["horror"],
+                    None,
+                    None,
+                ),
+            ],
+        )
+        .await
+        .unwrap();
+
+        let results = search_palette(&pool, "/lib", "author:stoker")
+            .await
+            .unwrap();
+        let titles: Vec<&str> = results.books.iter().map(|b| b.title.as_str()).collect();
+        assert!(
+            titles.contains(&"Dracula"),
+            "author:stoker should match Dracula, got {titles:?}"
+        );
+        assert!(
+            !titles.contains(&"Frankenstein"),
+            "author:stoker should not match Frankenstein"
+        );
+    }
+    #[tokio::test]
+    async fn palette_book_matches_series_facet() {
+        let _covers = CoversTempDir::new("palette_series_facet");
+        let pool = init_db("sqlite::memory:").await.unwrap();
+        replace_books(
+            &pool,
+            "/lib",
+            vec![
+                indexed(
+                    "a.epub",
+                    Some("Book One"),
+                    &["Author A"],
+                    &[],
+                    Some(("Dracula Chronicles", "1")),
+                    None,
+                ),
+                indexed("b.epub", Some("Unrelated"), &["Author B"], &[], None, None),
+            ],
+        )
+        .await
+        .unwrap();
+
+        let results = search_palette(&pool, "/lib", "series:dracula")
+            .await
+            .unwrap();
+        let titles: Vec<&str> = results.books.iter().map(|b| b.title.as_str()).collect();
+        assert!(
+            titles.contains(&"Book One"),
+            "series:dracula should match Book One, got {titles:?}"
+        );
+        assert!(
+            !titles.contains(&"Unrelated"),
+            "series:dracula should not match Unrelated"
+        );
+    }
+    #[tokio::test]
+    async fn palette_empty_query_returns_empty() {
+        let pool = init_db("sqlite::memory:").await.unwrap();
+        let results = search_palette(&pool, "/lib", "   ").await.unwrap();
+        assert!(results.books.is_empty());
+        assert!(results.authors.is_empty());
+        assert!(results.series.is_empty());
+        assert!(results.tags.is_empty());
+        assert_eq!(results.query, "");
+    }
+    #[tokio::test]
+    async fn palette_no_results() {
+        let _covers = CoversTempDir::new("palette_no_results");
+        let pool = init_db("sqlite::memory:").await.unwrap();
+        replace_books(
+            &pool,
+            "/lib",
+            vec![indexed("a.epub", Some("A"), &["Author"], &[], None, None)],
+        )
+        .await
+        .unwrap();
+
+        let results = search_palette(&pool, "/lib", "zzzznonexistent")
+            .await
+            .unwrap();
+        assert!(results.books.is_empty());
+        assert!(results.authors.is_empty());
+        assert!(results.series.is_empty());
+        assert!(results.tags.is_empty());
+    }
+    #[tokio::test]
+    async fn palette_scoped_to_library() {
+        let _covers = CoversTempDir::new("palette_scoped");
+        let pool = init_db("sqlite::memory:").await.unwrap();
+        replace_books(
+            &pool,
+            "/lib-a",
+            vec![indexed(
+                "a.epub",
+                Some("Alpha"),
+                &["Tolkien"],
+                &[],
+                None,
+                None,
+            )],
+        )
+        .await
+        .unwrap();
+        replace_books(
+            &pool,
+            "/lib-b",
+            vec![indexed(
+                "b.epub",
+                Some("Beta"),
+                &["Tolkien"],
+                &[],
+                None,
+                None,
+            )],
+        )
+        .await
+        .unwrap();
+
+        let results = search_palette(&pool, "/lib-a", "tolkien").await.unwrap();
+        // Books should only include lib-a
+        assert_eq!(results.books.len(), 1);
+        assert_eq!(results.books[0].title, "Alpha");
+        // Author book_count should be 1 (scoped to lib-a), not 2
+        assert_eq!(results.authors[0].book_count, 1);
+    }
+    #[tokio::test]
+    async fn palette_duration_populated() {
+        let _covers = CoversTempDir::new("palette_duration");
+        let pool = init_db("sqlite::memory:").await.unwrap();
+        replace_books(
+            &pool,
+            "/lib",
+            vec![indexed("a.epub", Some("A"), &["Author"], &[], None, None)],
+        )
+        .await
+        .unwrap();
+
+        let results = search_palette(&pool, "/lib", "author").await.unwrap();
+        // duration_ms should be populated (at least 0 — we just check it's set)
+        assert!(results.duration_ms < 10000, "duration should be reasonable");
+    }
+    /// #127 regression coverage: after collapsing the correlated
+    /// `book_count` / `EXISTS` subqueries into a single JOIN+GROUP BY,
+    /// `l.path = ?1` must still be applied **before** the aggregate so the
+    /// scoped library's count doesn't pick up rows from sibling libraries.
+    /// This exercises all three taxonomies (authors, series, tags) plus
+    /// ordering — the seeded set has 3 matching books in /lib-a and 2 in
+    /// /lib-b for the same author/series/tag, and the rare "Sole" author
+    /// only appears in /lib-b, so it must be absent from /lib-a results.
+    #[tokio::test]
+    async fn palette_taxonomy_counts_scoped_per_library() {
+        let _covers = CoversTempDir::new("palette_taxonomy_scoped");
+        let pool = init_db("sqlite::memory:").await.unwrap();
+
+        replace_books(
+            &pool,
+            "/lib-a",
+            vec![
+                indexed(
+                    "a1.epub",
+                    Some("Alpha One"),
+                    &["Shared Author"],
+                    &["Shared Tag"],
+                    Some(("Shared Series", "1")),
+                    None,
+                ),
+                indexed(
+                    "a2.epub",
+                    Some("Alpha Two"),
+                    &["Shared Author"],
+                    &["Shared Tag"],
+                    Some(("Shared Series", "2")),
+                    None,
+                ),
+                indexed(
+                    "a3.epub",
+                    Some("Alpha Three"),
+                    &["Shared Author"],
+                    &["Shared Tag"],
+                    Some(("Shared Series", "3")),
+                    None,
+                ),
+            ],
+        )
+        .await
+        .unwrap();
+        replace_books(
+            &pool,
+            "/lib-b",
+            vec![
+                indexed(
+                    "b1.epub",
+                    Some("Beta One"),
+                    &["Shared Author", "Sole Author"],
+                    &["Shared Tag"],
+                    Some(("Shared Series", "1")),
+                    None,
+                ),
+                indexed(
+                    "b2.epub",
+                    Some("Beta Two"),
+                    &["Shared Author"],
+                    &["Shared Tag"],
+                    Some(("Shared Series", "2")),
+                    None,
+                ),
+            ],
+        )
+        .await
+        .unwrap();
+
+        // Scope to /lib-a — author/series/tag counts must be 3, not 5.
+        let results = search_palette(&pool, "/lib-a", "Shared").await.unwrap();
+
+        let author = results
+            .authors
+            .iter()
+            .find(|a| a.name == "Shared Author")
+            .expect("Shared Author present in /lib-a results");
+        assert_eq!(
+            author.book_count, 3,
+            "author count must be scoped to /lib-a, got {results:?}"
+        );
+        assert!(
+            !results.authors.iter().any(|a| a.name == "Sole Author"),
+            "Sole Author lives only in /lib-b and must not appear"
+        );
+
+        let series = results
+            .series
+            .iter()
+            .find(|s| s.name == "Shared Series")
+            .expect("Shared Series present in /lib-a results");
+        assert_eq!(
+            series.book_count, 3,
+            "series count must be scoped to /lib-a"
+        );
+
+        let tag = results
+            .tags
+            .iter()
+            .find(|t| t.name == "Shared Tag")
+            .expect("Shared Tag present in /lib-a results");
+        assert_eq!(tag.book_count, 3, "tag count must be scoped to /lib-a");
+
+        // Cross-check /lib-b counts to make sure the same query returns 2.
+        let results_b = search_palette(&pool, "/lib-b", "Shared").await.unwrap();
+        let author_b = results_b
+            .authors
+            .iter()
+            .find(|a| a.name == "Shared Author")
+            .expect("Shared Author present in /lib-b results");
+        assert_eq!(author_b.book_count, 2);
+        let series_b = results_b
+            .series
+            .iter()
+            .find(|s| s.name == "Shared Series")
+            .expect("Shared Series present in /lib-b results");
+        assert_eq!(series_b.book_count, 2);
+        let tag_b = results_b
+            .tags
+            .iter()
+            .find(|t| t.name == "Shared Tag")
+            .expect("Shared Tag present in /lib-b results");
+        assert_eq!(tag_b.book_count, 2);
+    }
+    #[tokio::test]
+    async fn palette_author_count_reflects_overrides() {
+        // F5.1: the palette author count must match the merged
+        // (override-aware) view, not the raw `books_authors_link` count.
+        // Repro of the "Sanderson, Brandon still says 4 books" report:
+        // every canonical book for an author was reassigned to a
+        // differently-named author through the metadata edit form, so the
+        // palette must report 0 books for the source name and the full
+        // count for the destination name.
+        let _covers = CoversTempDir::new("palette_author_count_overrides");
+        let pool = init_db("sqlite::memory:").await.unwrap();
+        let user_id = crate::auth::create_user(&pool, "admin", "securepassword1")
+            .await
+            .unwrap()
+            .id;
+
+        // Two books canonically by "Last, First", plus one book by the
+        // already-correct "First Last" so the destination author has a
+        // canonical anchor (palette visibility requires ≥1 canonical link).
+        replace_books(
+            &pool,
+            "/lib",
+            vec![
+                indexed("a.epub", Some("A"), &["Last, First"], &[], None, None),
+                indexed("b.epub", Some("B"), &["Last, First"], &[], None, None),
+                indexed("c.epub", Some("C"), &["First Last"], &[], None, None),
+            ],
+        )
+        .await
+        .unwrap();
+
+        // User edits a.epub and b.epub through the metadata form to
+        // rename their author to "First Last" — overrides only, no
+        // change to the relational link table.
+        let books = list_books(&pool, "/lib").await.unwrap();
+        for filename in ["a.epub", "b.epub"] {
+            let book = books.iter().find(|b| b.filename == filename).unwrap();
+            let uuid = book.unique_identifier.clone().unwrap();
+            let ov = MetadataOverrides {
+                creators: Some(vec![Contributor {
+                    name: "First Last".into(),
+                    role: Some("aut".into()),
+                    file_as: None,
+                    id: None,
+                }]),
+                ..Default::default()
+            };
+            upsert_metadata_overrides(&pool, &uuid, &ov, false, user_id)
+                .await
+                .unwrap();
+        }
+
+        let results = search_palette(&pool, "/lib", "Last").await.unwrap();
+
+        // Source author still visible (canonical anchor remains), but
+        // count must reflect the effective view: 0 books.
+        let source = results
+            .authors
+            .iter()
+            .find(|a| a.name == "Last, First")
+            .expect("source author still appears in palette");
+        assert_eq!(
+            source.book_count, 0,
+            "renamed-away author must report effective count 0, got {results:?}",
+        );
+
+        // Destination author picks up the override-renamed books on top
+        // of its own canonical anchor: 1 + 2 = 3.
+        let dest = results
+            .authors
+            .iter()
+            .find(|a| a.name == "First Last")
+            .expect("destination author present");
+        assert_eq!(
+            dest.book_count, 3,
+            "destination author must include override-renamed books, got {results:?}",
+        );
+    }
+    #[tokio::test]
+    async fn palette_tag_count_reflects_overrides() {
+        // F5.1: same shape for tags. `overrides.subjects` replaces the
+        // canonical tag list wholesale, so a book moved between tags
+        // must shift both counts.
+        let _covers = CoversTempDir::new("palette_tag_count_overrides");
+        let pool = init_db("sqlite::memory:").await.unwrap();
+        let user_id = crate::auth::create_user(&pool, "admin", "securepassword1")
+            .await
+            .unwrap()
+            .id;
+
+        replace_books(
+            &pool,
+            "/lib",
+            vec![
+                indexed("a.epub", Some("A"), &["X"], &["tag-source"], None, None),
+                indexed("b.epub", Some("B"), &["X"], &["tag-source"], None, None),
+                indexed("c.epub", Some("C"), &["X"], &["tag-dest"], None, None),
+            ],
+        )
+        .await
+        .unwrap();
+
+        // Move a.epub off tag-source and onto tag-dest via override.
+        let books = list_books(&pool, "/lib").await.unwrap();
+        let a = books.iter().find(|b| b.filename == "a.epub").unwrap();
+        let uuid = a.unique_identifier.clone().unwrap();
+        let ov = MetadataOverrides {
+            subjects: Some(vec!["tag-dest".into()]),
+            ..Default::default()
+        };
+        upsert_metadata_overrides(&pool, &uuid, &ov, false, user_id)
+            .await
+            .unwrap();
+
+        let results = search_palette(&pool, "/lib", "tag-").await.unwrap();
+        let source = results
+            .tags
+            .iter()
+            .find(|t| t.name == "tag-source")
+            .expect("tag-source still visible (canonical anchor remains)");
+        assert_eq!(
+            source.book_count, 1,
+            "tag-source should drop a.epub after override, got {results:?}",
+        );
+        let dest = results
+            .tags
+            .iter()
+            .find(|t| t.name == "tag-dest")
+            .expect("tag-dest present");
+        assert_eq!(
+            dest.book_count, 2,
+            "tag-dest should add the override-tagged a.epub, got {results:?}",
+        );
+    }
+    #[tokio::test]
+    async fn palette_series_count_reflects_overrides() {
+        // F5.1: same shape as palette_author_count_reflects_overrides
+        // but for the series tile. Books moved into a series via
+        // `overrides.series` must add to the destination count; books
+        // moved out drop from the source count.
+        let _covers = CoversTempDir::new("palette_series_count_overrides");
+        let pool = init_db("sqlite::memory:").await.unwrap();
+        let user_id = crate::auth::create_user(&pool, "admin", "securepassword1")
+            .await
+            .unwrap()
+            .id;
+
+        replace_books(
+            &pool,
+            "/lib",
+            vec![
+                indexed(
+                    "a.epub",
+                    Some("A"),
+                    &["X"],
+                    &[],
+                    Some(("Series Source", "1")),
+                    None,
+                ),
+                indexed(
+                    "b.epub",
+                    Some("B"),
+                    &["X"],
+                    &[],
+                    Some(("Series Source", "2")),
+                    None,
+                ),
+                indexed(
+                    "c.epub",
+                    Some("C"),
+                    &["X"],
+                    &[],
+                    Some(("Series Dest", "1")),
+                    None,
+                ),
+            ],
+        )
+        .await
+        .unwrap();
+
+        // Move a.epub from Series Source to Series Dest via override.
+        let books = list_books(&pool, "/lib").await.unwrap();
+        let a = books.iter().find(|b| b.filename == "a.epub").unwrap();
+        let uuid = a.unique_identifier.clone().unwrap();
+        let ov = MetadataOverrides {
+            series: Some("Series Dest".into()),
+            ..Default::default()
+        };
+        upsert_metadata_overrides(&pool, &uuid, &ov, false, user_id)
+            .await
+            .unwrap();
+
+        // "Series" matches both names.
+        let results = search_palette(&pool, "/lib", "Series").await.unwrap();
+        let source = results
+            .series
+            .iter()
+            .find(|s| s.name == "Series Source")
+            .expect("Series Source still visible (canonical anchor remains)");
+        assert_eq!(
+            source.book_count, 1,
+            "Series Source should count only b.epub after a.epub is overridden away, got {results:?}",
+        );
+        let dest = results
+            .series
+            .iter()
+            .find(|s| s.name == "Series Dest")
+            .expect("Series Dest present");
+        assert_eq!(
+            dest.book_count, 2,
+            "Series Dest should count its canonical c.epub plus the override-moved a.epub, got {results:?}",
+        );
+    }
+    #[tokio::test]
+    async fn palette_series_author_display_reflects_override() {
+        // F5.1: the "by X" line on a series tile must follow the first
+        // book's effective creator, not the canonical one — otherwise
+        // renaming the author through the metadata edit form leaves the
+        // palette showing the old name.
+        let _covers = CoversTempDir::new("palette_series_author_display");
+        let pool = init_db("sqlite::memory:").await.unwrap();
+        let user_id = crate::auth::create_user(&pool, "admin", "securepassword1")
+            .await
+            .unwrap()
+            .id;
+
+        replace_books(
+            &pool,
+            "/lib",
+            vec![indexed(
+                "k1.epub",
+                Some("K1"),
+                &["Old Name"],
+                &[],
+                Some(("Kingsway", "1")),
+                None,
+            )],
+        )
+        .await
+        .unwrap();
+
+        let books = list_books(&pool, "/lib").await.unwrap();
+        let uuid = books[0].unique_identifier.clone().unwrap();
+
+        let ov = MetadataOverrides {
+            creators: Some(vec![Contributor {
+                name: "New Name".into(),
+                role: Some("aut".into()),
+                file_as: None,
+                id: None,
+            }]),
+            ..Default::default()
+        };
+        upsert_metadata_overrides(&pool, &uuid, &ov, false, user_id)
+            .await
+            .unwrap();
+
+        let results = search_palette(&pool, "/lib", "Kingsway").await.unwrap();
+        let kingsway = results
+            .series
+            .iter()
+            .find(|s| s.name == "Kingsway")
+            .expect("Kingsway present");
+        assert_eq!(
+            kingsway.author_display.as_deref(),
+            Some("New Name"),
+            "palette author line must follow override.creators, got {results:?}",
+        );
+    }
+    /// #127: capture `EXPLAIN QUERY PLAN` for each of the three rewritten
+    /// taxonomy queries and assert the planner uses the link-table indexes.
+    /// This is a structural check — it doesn't pin the literal plan string
+    /// (SQLite's wording can shift across point releases) but it does fail
+    /// loudly if any of the link tables fall back to a full SCAN, which
+    /// would defeat the whole point of this optimization.
+    #[tokio::test]
+    async fn palette_taxonomy_query_plans_use_indexes() {
+        let pool = init_db("sqlite::memory:").await.unwrap();
+
+        async fn plan_text(pool: &SqlitePool, sql: &str) -> String {
+            let rows = sqlx::query(&format!("EXPLAIN QUERY PLAN {sql}"))
+                .bind("/lib")
+                .bind("%x%")
+                .bind(5_i32)
+                .fetch_all(pool)
+                .await
+                .unwrap();
+            rows.iter()
+                .map(|r| r.get::<String, _>("detail"))
+                .collect::<Vec<_>>()
+                .join("\n")
+        }
+
+        // Authors — override-aware count, must still drive through the
+        // covering `books_authors_link` index when checking visibility
+        // and when the no-override branch of the CASE runs.
+        let plan = plan_text(
+            &pool,
+            "SELECT a.id, a.name, \
+              (SELECT COUNT(*) FROM books b \
+                 JOIN libraries l2 ON l2.id = b.library_id \
+                 LEFT JOIN metadata_overrides mo ON mo.book_uuid = b.uuid \
+                WHERE l2.path = ?1 \
+                  AND CASE \
+                        WHEN mo.book_uuid IS NOT NULL \
+                             AND json_type(mo.overrides, '$.creators') IS NOT NULL \
+                          THEN EXISTS (SELECT 1 FROM json_each(mo.overrides, '$.creators') je \
+                                        WHERE json_extract(je.value, '$.name') = a.name) \
+                        ELSE EXISTS (SELECT 1 FROM books_authors_link bal \
+                                      WHERE bal.book = b.id AND bal.author = a.id) \
+                      END \
+              ) AS book_count \
+             FROM authors a \
+             WHERE a.name LIKE ?2 ESCAPE '\\' \
+               AND EXISTS (SELECT 1 FROM books_authors_link bal \
+                             JOIN books b ON b.id = bal.book \
+                             JOIN libraries l ON l.id = b.library_id \
+                            WHERE bal.author = a.id AND l.path = ?1) \
+             ORDER BY book_count DESC, a.name \
+             LIMIT ?3",
+        )
+        .await;
+        assert!(
+            !plan.contains("SCAN books_authors_link") && !plan.contains("SCAN bal"),
+            "authors plan should not full-scan the link table:\n{plan}"
+        );
+
+        // Series — override-aware count + author_display, must still drive
+        // through the `books_series_link` index for visibility and the
+        // no-override branch of the CASE.
+        let plan = plan_text(
+            &pool,
+            "SELECT s.id, s.name, \
+              (SELECT COUNT(*) FROM books b \
+                 JOIN libraries l2 ON l2.id = b.library_id \
+                 LEFT JOIN metadata_overrides mo ON mo.book_uuid = b.uuid \
+                WHERE l2.path = ?1 \
+                  AND CASE \
+                        WHEN mo.book_uuid IS NOT NULL \
+                             AND json_type(mo.overrides, '$.series') IS NOT NULL \
+                          THEN json_extract(mo.overrides, '$.series') = s.name \
+                        ELSE EXISTS (SELECT 1 FROM books_series_link bsl \
+                                      WHERE bsl.book = b.id AND bsl.series = s.id) \
+                      END \
+              ) AS book_count \
+             FROM series s \
+             WHERE s.name LIKE ?2 ESCAPE '\\' \
+               AND EXISTS (SELECT 1 FROM books_series_link bsl \
+                             JOIN books b ON b.id = bsl.book \
+                             JOIN libraries l ON l.id = b.library_id \
+                            WHERE bsl.series = s.id AND l.path = ?1) \
+             ORDER BY book_count DESC, s.name \
+             LIMIT ?3",
+        )
+        .await;
+        assert!(
+            !plan.contains("SCAN books_series_link") && !plan.contains("SCAN bsl"),
+            "series plan should not full-scan the link table:\n{plan}"
+        );
+
+        // Tags — override-aware count, must still drive through the
+        // `books_tags_link` index for visibility and the no-override
+        // branch of the CASE.
+        let plan = plan_text(
+            &pool,
+            "SELECT t.id, t.name, \
+              (SELECT COUNT(*) FROM books b \
+                 JOIN libraries l2 ON l2.id = b.library_id \
+                 LEFT JOIN metadata_overrides mo ON mo.book_uuid = b.uuid \
+                WHERE l2.path = ?1 \
+                  AND CASE \
+                        WHEN mo.book_uuid IS NOT NULL \
+                             AND json_type(mo.overrides, '$.subjects') IS NOT NULL \
+                          THEN EXISTS (SELECT 1 FROM json_each(mo.overrides, '$.subjects') je \
+                                        WHERE je.value = t.name) \
+                        ELSE EXISTS (SELECT 1 FROM books_tags_link btl \
+                                      WHERE btl.book = b.id AND btl.tag = t.id) \
+                      END \
+              ) AS book_count \
+             FROM tags t \
+             WHERE t.name LIKE ?2 ESCAPE '\\' \
+               AND EXISTS (SELECT 1 FROM books_tags_link btl \
+                             JOIN books b ON b.id = btl.book \
+                             JOIN libraries l ON l.id = b.library_id \
+                            WHERE btl.tag = t.id AND l.path = ?1) \
+             ORDER BY book_count DESC, t.name \
+             LIMIT ?3",
+        )
+        .await;
+        assert!(
+            !plan.contains("SCAN books_tags_link") && !plan.contains("SCAN btl"),
+            "tags plan should not full-scan the link table:\n{plan}"
+        );
+    }
+}
